# macho: fix performance regression for x86_64 and -dead_strip