Prior to this patch, we would unnecessarily spin in `markLive` routine when dead stripping due to inability to handle target atoms described by hard-coded offsets wrt to the source relocatable object file (i.e., `r_extern == 0` which normally occurs `X86_64_RELOC_UNSIGNED` and `X86_64_RELOC_SIGNED`). This is now fixed by caching sorted source symbols by address per object and finding the containing atom of this type of relocation, and marking it live.

With this patch, macOS users on `x86_64` can now experience comparable performance as `aarch64` users. Some numbers run against standalone `zld` linking `stage3`. The link time of `stage3` with dead stripping has now been reduced by 1 second making `zld` just 1.40x slower than `lld` and faster than `ld64`:

```
❯ hyperfine ./zld ./lld ./ld64 --warmup 5
Benchmark 1: ./zld
  Time (mean ± σ):      2.300 s ±  0.013 s    [User: 1.484 s, System: 0.797 s]
  Range (min … max):    2.284 s …  2.321 s    10 runs
 
Benchmark 2: ./lld
  Time (mean ± σ):      1.639 s ±  0.010 s    [User: 1.655 s, System: 0.624 s]
  Range (min … max):    1.627 s …  1.656 s    10 runs
 
Benchmark 3: ./ld64
  Time (mean ± σ):      2.682 s ±  0.060 s    [User: 3.638 s, System: 0.721 s]
  Range (min … max):    2.627 s …  2.802 s    10 runs
 
Summary
  './lld' ran
    1.40 ± 0.01 times faster than './zld'
    1.64 ± 0.04 times faster than './ld64'
```